Time: 1 ms (97.44%), Space: 39.3 MB (79.65%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -2,22 +2,21 @@ class Solution {
     public int mySqrt(int x) {
         if(x==0) return 0;
     
-        long l=0;
-        long r=x;
+        int l=0;
+        int r=x;
         int ans=0;
         while(l<=r){
             long mid=(l+r)/2;
             long sqr=mid*mid;
             if(sqr==x){
                 return (int)mid;
-            }else if(sqr<x){
-                l=mid+1;
+            }else if( sqr < x){
+                l=(int)mid+1;
                 ans=(int)mid;
             }else{
-                r=mid-1;
+                r=(int)mid-1;
             }   
         }
         return ans;
-
     }
 }
